// src/app/Common/types.tsx
export interface UserData {
  firstName: string;
  image: string;
}

export interface Product {
  id: number;
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  brand: string;
  category: string;
  thumbnail: string;
  images: string[];
}

export interface ProductsData {
  products: Product[];
  total?: number;
  skip?: number;
  limit?: number;
}